Remove even more code
